Python Annotations

+++ 0x00-python_variable_annotations/2-floor.py
@@ -3,6 +3,6 @@ import math
 """Floor function with annotations"""
 
 
-def floor(num: float) -> float:
+def floor(num: float) -> int:
     """Return the floor of a number."""
     return math.floor(num)
